Update objectives.R
genotype.factors() creates factors for cons$factors. Usually this is done using C++ code and returns a vector. However, under the simple prosecution hypothesis of nUnknowns=0 and doDropin=F, factors is created in R and returns a matrix of 1 column. This is caused by het starting its life inside genotype.factors() as a 1 column matrix. This change simply creates het as a vector instead.