Updated multiRun and optimization.params
Allowed multiRun to handle the user defining fixed arguments. Moved the creation of the objective function outside of the for loop of multiRun so that the computationally expensive step (genotype combinations) is only done once. Added objective parameter to optimization.params so that you can compute the objective function beforehand, and pass it to optimization.params.