pack.admin.input tests
Removed the argument option to checkFiles, safer and simpler if it is
always done. Build tests then failed. It looks like the build test was
never doing its job as the file paths to the input files were wrong, and
a typo in the allele database file, yet the build tests didn't
previously throw an error.

Few other minor bits