Allowed handling of dropin values that are too high
Before making objective function we find the maximum allele frequency using getMaxAF(). This is then incorporated into the objective function, where we check that af*(dropin*(1-dropout)) is not greater than 1 before running. If so we return a likelihood of 0 or -Inf (depending on if we want a log returned).