Vignette and test run on 2 cores max.
CRAN requires that code runs on two cores or less so as to preserve their computational resources.
There are now two functions to control openmp, with suitably obscure names (e.g. for experts only).
These functions are called at the start of the vignette and at setup and teardown of each test.

Closes #13